MC-34514: Acl resource changes
