+++ InventoryAdminUi/Controller/Adminhtml/Source/Save.php
@@ -29,7 +29,7 @@ class Save extends Action implements HttpPostActionInterface
     /**
      * @see _isAllowed()
      */
-    const ADMIN_RESOURCE = 'Magento_InventoryApi::source';
+    const ADMIN_RESOURCE = 'Magento_InventoryApi::source_edit';
 
     /**
      * @var SourceInterfaceFactory
