+++ InventoryAdminUi/Test/Integration/_files/user_with_restricted_role.php
@@ -18,7 +18,7 @@ use Magento\Authorization\Model\Rules;
 //Creating a new admin user with a custom role to safely change role settings without affecting the main user's role.
 /** @var Role $role */
 $role = Bootstrap::getObjectManager()->get(RoleFactory::class)->create();
-$role->setName('test_custom_role');
+$role->setName('source_access_role');
 $role->setData('role_name', $role->getName());
 $role->setRoleType(Group::ROLE_TYPE);
 $role->setUserType((string)UserContextInterface::USER_TYPE_ADMIN);
@@ -40,7 +40,7 @@ $rules->saveRel();
 $user = Bootstrap::getObjectManager()->create(User::class);
 $user->setFirstname("John")
     ->setLastname("Doe")
-    ->setUsername('customRoleUser')
+    ->setUsername('sourceAccessUser')
     ->setPassword(\Magento\TestFramework\Bootstrap::ADMIN_PASSWORD)
     ->setEmail('adminUser@example.com')
     ->setIsActive(1)
