+++ InventoryAdminUi/Test/Integration/_files/user_with_restricted_role_rollback.php
@@ -19,7 +19,7 @@ use Magento\Authorization\Model\Rules;
 $objectManager = Bootstrap::getObjectManager();
 /** @var User $user */
 $user = $objectManager->create(UserFactory::class)->create();
-$user->load('customRoleUser', 'username');
+$user->load('sourceAccessUser', 'username');
 
 if ($user->getId() !== null) {
     $user->delete();
@@ -27,7 +27,7 @@ if ($user->getId() !== null) {
 
 /** @var Role $role */
 $role = Bootstrap::getObjectManager()->get(RoleFactory::class)->create();
-$role->load('test_custom_role', 'role_name');
+$role->load('source_access_role', 'role_name');
 /** @var Rules $rules */
 $rules = Bootstrap::getObjectManager()->get(RulesFactory::class)->create();
 $rules->load($role->getId(), 'role_id');
