+++ InventoryAdminUi/Ui/Component/Listing/Column/EditAction.php
@@ -1,0 +1,65 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Magento\InventoryAdminUi\Ui\Component\Listing\Column;
+
+use Magento\Framework\App\ObjectManager;
+use Magento\Framework\AuthorizationInterface;
+use Magento\Framework\UrlInterface;
+use Magento\Framework\View\Element\UiComponent\ContextInterface;
+use Magento\Framework\View\Element\UiComponentFactory;
+
+/**
+ * Represents Edit link in sources grid for entity by its identifier field
+ */
+class EditAction extends \Magento\Backend\Ui\Component\Listing\Column\EditAction
+{
+    /**
+     * @var AuthorizationInterface
+     */
+    private $authorization;
+
+    /**
+     * @param ContextInterface $context
+     * @param UiComponentFactory $uiComponentFactory
+     * @param UrlInterface $urlBuilder
+     * @param array $components
+     * @param array $data
+     * @param AuthorizationInterface|null $authorization
+     */
+    public function __construct(
+        ContextInterface $context,
+        UiComponentFactory $uiComponentFactory,
+        UrlInterface $urlBuilder,
+        array $components = [],
+        array $data = [],
+        ?AuthorizationInterface $authorization = null
+    ) {
+        $this->authorization = $authorization ?? ObjectManager::getInstance()->get(AuthorizationInterface::class);
+        parent::__construct($context, $uiComponentFactory, $urlBuilder, $components, $data);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function prepareDataSource(array $dataSource)
+    {
+        $dataSource = parent::prepareDataSource($dataSource);
+        $actionsName = $this->getData('name');
+
+        if (isset($dataSource['data']['items'])) {
+            foreach ($dataSource['data']['items'] as &$item) {
+                if (!empty($item[$actionsName]['edit'])) {
+                    $item[$actionsName]['edit']['hidden'] =
+                        !$this->authorization->isAllowed('Magento_InventoryApi::source_edit');
+                }
+            }
+        }
+
+        return $dataSource;
+    }
+}
